Reuse search page across website and editor
